+++ src/main/java/ru/geekbrains/java/oop/at/lesson1/CatFood.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.java.oop.at;
+package ru.geekbrains.java.oop.at.lesson1;
 
 public class CatFood {
     private String productName;
