+++ src/main/java/ru/geekbrains/java/oop/at/lesson1/obstacles/Obstacles.java
@@ -1,0 +1,7 @@
+package ru.geekbrains.java.oop.at.lesson1.obstacles;
+
+import ru.geekbrains.java.oop.at.lesson1.participants.Runner;
+
+public interface Obstacles {
+    void check(Runner runner);
+}
