+++ src/main/java/ru/geekbrains/java/oop/at/lesson1/obstacles/Treadmill.java
@@ -1,6 +1,6 @@
-package ru.geekbrains.java.oop.at.obstacles;
+package ru.geekbrains.java.oop.at.lesson1.obstacles;
 
-import ru.geekbrains.java.oop.at.participants.Runner;
+import ru.geekbrains.java.oop.at.lesson1.participants.Runner;
 
 public class Treadmill implements Obstacles {
     private int length;
