+++ src/main/java/ru/geekbrains/java/oop/at/lesson1/obstacles/Wall.java
@@ -1,6 +1,6 @@
-package ru.geekbrains.java.oop.at.obstacles;
+package ru.geekbrains.java.oop.at.lesson1.obstacles;
 
-import ru.geekbrains.java.oop.at.participants.Runner;
+import ru.geekbrains.java.oop.at.lesson1.participants.Runner;
 
 public class Wall implements Obstacles {
     private int width;
