--- src/main/java/ru/geekbrains/java/oop/at/obstacles/Obstacles.java
@@ -1,7 +1,0 @@
-package ru.geekbrains.java.oop.at.obstacles;
-
-import ru.geekbrains.java.oop.at.participants.Runner;
-
-public interface Obstacles {
-    void check(Runner runner);
-}
